video-api/requirements.txt
requests
moviepy
pillow
nltk
python-dotenv
pyttsx3
matplotlib
networkx
